bossqu-ai/pyproject.toml
[tool.poetry]
name = "bossqu-ai"
version = "0.0.1"
description = ""
authors = ["Ediyono <ediyono90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
fastapi = "0.115.9"
python = ">=3.10,<4.0"
langchain = "^0.3.25"
langchain-core = "^0.3.58"
langchain-openai = "^0.3.16"
langchain-community = "^0.3.23"
sentence-transformers = "^4.1.0"
chromadb = "^1.0.7"
protobuf = "3.20.3"
langchain-huggingface = "^0.1.2"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.34.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
